split out monolithic test file
